📝 README: Remove unneccesary installation step for ocrd_calamari etc.
There is a part in the README that suggests that a seperate installation
step for ocrd_calamari was necessary. I do believe that was because of
issues with the varying TF versions or some other problems unrelated to
ocrd_butler.

ocrd_calamari runs fine with ocrd_all's installation method (make all),
so I am proposing to remove this part of the README.