[#13030] ensure workflow processors have output_file_grp values
workflow processors without explicitly set `output_file_grp` values now
get one based on the processor name and its position in the workflow
assigned at time of creation. Default value picked from the processor
executable's option `-J` output is being ignored, as it is not guaranteed
to always be the perfect fit in all scenarios.
The only way to enforce a specific `output_file_grp` default value for a
processor's invocations with that field missing is now via the
`ocrd_butler.config.Config.PROCESSOR_SETTINGS` structure (as a string
value).